boilerplate/requirements.txt
attrs==19.3.0
importlib-metadata==1.4.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.1.0
pyrsistent==0.15.7
PyYAML==5.3
six==1.13.0
zipp==0.6.0
